requirements.txt
mysql.connector
requests
colorama
pyfiglet
python-dotenv
